Revert "Fix error when running on logon"
This reverts commit b725600f85a2b61c4b72142ef2f0665f7dcb6dff.

This broke the basic functionality of the window switcher.

I thought I had tested this, but I hadn't actually, because I had
extracted the script from my dotfiles into its own repo, but I was still
running the old version, separate from the version I was editing and
wanting to test. The auto-reload feature I built into the script was
particularly misleading in this case, because it reloaded the script
(with a little splash message and everything) upon saving changes to
the unrelated file, as it's based on seeing the file name in a window
title, not watching the file system.