Retry fix for error when running on logon (now with scoping fix)
When the script was set to run at startup (that is, at logon),
it tried to initialize the TaskbarList object too early, presumably
before the taskbar was ready, and got this error:

Error: (0x80004001) Not implemented

	040: ITaskbarList_VTable := { HrInit: 3, AddTab: 4, DeleteTab: 5, ActivateTab: 6, SetActiveAlt: 7, }
	048: TaskbarList := ComObject(CLSID_TaskbarList, IID_ITaskbarList)
▶	049: ComCall(ITaskbarList_VTable.HrInit, TaskbarList)
	051: TempHiddenWindows := []
	054: {

The fix is to delay initialization (`HrInit`) until needed.

This fix was previously attempted in b725600f85a2b61c4b72142ef2f0665f7dcb6dff
and reverted in 1fe9a8f3c375746d30b641190322be6f0d04dd9f
because of a tiny variable scoping issue that broke the whole script.
Without `global TaskbarListInitialized`, since it was assigned within
the function, it created a separate local, shadowing the global,
presumably leading to it calling `HrInit` multiple times (even for a
single showing of the window switcher).