Fix memory issue in `FileGetVersionInfo_AW`
This only seemed to affect the compiled `app-switcher.exe`,
although the problem may have merely been more hidden when running
as an uncompiled script.

This sometimes manifested as an error message:

    Critical Error: Invalid memory read/write.

    ▶	540: Apps.Push({ Icon: GetAppIconHandle(Window), Title: Title, HWND: Window, })

    The program is now unstable and will exit.

    Call stack:
    *#1 (540) : [Array.Prototype.Push] Apps.Push({ Icon: GetAppIconHandle(Window), Title: Title, HWND: Window, })
    *#1 (540) : [<Hotkey>] Apps.Push({ Icon: GetAppIconHandle(Window), Title: Title, HWND: Window, })
    > #Tab

But usually app-switcher.exe just crashed silently when pressing Win+Tab

In no case did it give useful info for debugging this other than that it
was a memory issue. I identified the problem by disabling parts of the
script that were not shared with the working `window-switcher.exe`,
namely `#Include`, `FileInstall`, and `DllCall`, then re-enabling things
once I found that it seemed related to `DllCall`, and finally asking
ChatGPT to identify memory issues in `FileGetVersionInfo_AW`, since it
was originally written for AHK v1 and I ported it myself to AHK v2, and
it didn't go super smoothly as far as I recall.